Actualización 20/07/2024. Se ha establecido correctamente relacion entre usuarios y detalles. Falta poder meter imagenes y establecer relacion entre detalles y genero.

+++ src/test/java/com/eoi/Fruggy/EntidadesTest/DetalleTest.java
@@ -39,7 +39,7 @@ public class DetalleTest {
         //detalle.setDetalleGenero(genero);
 
         Detalle detalleGuardado = repoDetalle.save(detalle);
-        Optional <Detalle> detalleRecuperado= repoDetalle.findById((int) detalleGuardado.getId());
+        Optional <Detalle> detalleRecuperado= repoDetalle.findById((long) detalleGuardado.getId());
 
 
         assertNotNull(detalleRecuperado.get().getId());
